依存ライブラリの更新

+++ include/ggrpc/impl/client/client_reader_writer.h
@@ -449,7 +449,7 @@ class ClientReaderWriter {
         SPDLOG_DEBUG("gRPC Read finished");
       } else {
         SPDLOG_ERROR("gRPC error: {} ({})", grpc_status_.error_message(),
-                     grpc_status_.error_code());
+                     (int)grpc_status_.error_code());
         SPDLOG_ERROR("   details: {}", grpc_status_.error_details());
       }
       // 正常終了
